+++ projects/ngx-dmaster-ui/src/lib/components/buttons/copy-button/index.ts
@@ -1,0 +1,4 @@
+export * from './copy-button.component';
+export * from './copy-button.tokens';
+export * from './copy-button.types';
+export * from './copy-to-clipboard.directive';