+++ projects/ngx-test-app/src/app/core/i18n/translations.types.ts
@@ -99,6 +99,7 @@ export interface DashboardTranslations {
       emptyState: string;
       toggleGroup: string;
       buttonGroup: string;
+      copyButton: string;
       otp: string;
     };
   };
@@ -387,6 +388,7 @@ export interface DashboardTranslations {
       catalogue: string;
       current: string;
       theming: string;
+      copy: string;
     };
     focusItems: {
       freezeTitle: string;
@@ -454,6 +456,7 @@ export interface DashboardTranslations {
     articleNewer: string;
     articleOlder: string;
     articles: {
+      v09: BlogArticleTranslations;
       v08: BlogArticleTranslations;
       v07: BlogArticleTranslations;
       v06: BlogArticleTranslations;
@@ -507,6 +510,7 @@ export interface DashboardTranslations {
     emptyState: SimplePageTranslations;
     toggleGroup: SimplePageTranslations;
     buttonGroup: SimplePageTranslations;
+    copyButton: SimplePageTranslations;
     otp: SimplePageTranslations;
   };
 }